Change to int8 network input and buffer

+++ Core/Inc/ring_buffer.h
@@ -14,18 +14,18 @@
 #include <arm_math.h>
 
 struct RingBuffer {
-	float32_t data[BUFFERSIZE][N_MFCC];
+	int8_t data[BUFFERSIZE][N_MFCC];
 	uint16_t last_inference_head;
 	uint16_t buffer_ptr;
 };
 
 void init_ring_buffer(struct RingBuffer* rb);
 
-void insert_data(struct RingBuffer* rb, float32_t* data);
+void insert_data(struct RingBuffer* rb, int8_t* data);
 
 void increment_buffer_ptr(struct RingBuffer* rb);
 
-void copy_inference_batch(struct RingBuffer* rb, float32_t* batch);
+void copy_inference_batch(struct RingBuffer* rb, int8_t* batch);
 
 void update_last_inference_head(struct RingBuffer* rb);
